fix state + implement get season details

+++ providers/kinopub/kinopub.go
@@ -300,6 +300,6 @@ func StripImdbID(id string) int {
 	return i
 }
 
-func ToUID(id int64) string {
+func ToUID(id int) string {
 	return fmt.Sprintf("%s%d", providers.ID_TYPE_KINOHUB, id)
 }
